[spmd] rename _utils to common_utils
as titled, no need to review
