+++ test/spmd/tensor/test_megatron_example.py
@@ -3,7 +3,7 @@ import torch
 import torch.distributed as dist
 import functools
 from torch.testing._internal.common_utils import run_tests
-from spmd.test._utils import DistTensorTestBase, with_comms  # type: ignore
+from spmd.test.common_utils import DistTensorTestBase, with_comms  # type: ignore
 from spmd import distribute_tensor, DeviceMesh, DTensor, Shard, Replicate
