+++ test/spmd/tensor/test_redistribute.py
@@ -6,7 +6,7 @@ from torch.distributed.distributed_c10d import ReduceOp
 
 from torch.testing._internal.common_utils import run_tests
 
-from spmd.test._utils import DistTensorTestBase, with_comms  # type: ignore
+from spmd.test.common_utils import DistTensorTestBase, with_comms  # type: ignore
 from spmd.tensor import DeviceMesh, DTensor, Replicate, Shard, _Partial
 from spmd import distribute_tensor
